+++ filesystem/std/src/syscalls/fs.rs
@@ -1,15 +1,15 @@
 #[repr(packed)]
 #[derive(Debug)]
 pub struct Stat {
-    pub directory_name_length: usize,
-    pub directory_size: usize,
+    pub directory_name_length: u32,
+    pub directory_size: u32,
     pub is_folder: bool,
-    pub child_nodes_count: usize,
+    pub child_nodes_count: u32,
 }
 
 pub fn open(file_path: &str) -> usize {
     use syscalls::syscall::syscall2;
-    // SYSCALL(SYS_FOPEN, ptr, size)
+    // SYSCALL(FS_OPEN, ptr, size)
     unsafe {
         syscall2(0x1, file_path.as_ptr() as usize, file_path.len())
     }
@@ -17,14 +17,14 @@ pub fn open(file_path: &str) -> usize {
 
 pub fn read(fd: usize, buffer: &mut [u8]) -> usize {
     use syscalls::syscall::syscall3;
-    // SYSCALL(SYS_READ, fd, ptr, size)
+    // SYSCALL(FS_READ, fd, ptr, size)
     unsafe {
         syscall3(0x3, fd, buffer.as_ptr() as usize, buffer.len())
     }
 }
 
 pub fn stat(parent_directory: &str, child_node: usize) -> Stat {
-    // SYSCALL(SYS_STAT, ptr, size, stat_structure_ptr, child_node)
+    // SYSCALL(FS_STAT, ptr, size, stat_structure_ptr, child_node)
     use syscalls::syscall::syscall4;
 
     let mut stat = Stat {
@@ -41,7 +41,7 @@ pub fn stat(parent_directory: &str, child_node: usize) -> Stat {
 }
 
 pub fn read_name(parent_directory: &str, read_buffer: &mut [u8], child_node: usize) {
-    // SYSCALL(SYS_STAT, ptr, size, read_buffer_ptr, read_buffer_size, child_node)
+    // SYSCALL(FS_DIR_NAME, ptr, size, read_buffer_ptr, read_buffer_size, child_node)
     use syscalls::syscall::syscall5;
 
     unsafe {
